refactor: data type

+++ MamaFit.BusinessObjects/DTO/GhtkDto/SubmitOrder/OrderProductDto.cs
@@ -5,5 +5,5 @@ public class OrderProductDto
     public string? Name { get; set; }
     public decimal? Weight { get; set; }
     public int Quantity { get; set; }
-    public float? Price { get; set; }
+    public decimal? Price { get; set; }
 }
